==> notebooks/02_feature_engineering.ipynb <==
#import neccessary libraries
import pandas as pd 
from sklearn.preprocessing import LabelEncoder

#load the dataset
df=pd.read_csv(r"C:\Users\kumar\Academic_Decision_Support_System\data\synthetic\student_academic_data.csv")
df.head()

#checking null values
df.isnull().sum()

# handle missing subject scores
subject_cols = [
    "physics","chemistry","maths","biology",
    "english","computer","commerce_sub","arts_sub"
]

df[subject_cols] = df[subject_cols].fillna(0)

# encoding Categorical Variables

le = LabelEncoder()
df["stream_encoded"] = le.fit_transform(df["stream"])
df["family_support"] = df["family_support"].astype(int)
df["school_support"] = df["school_support"].astype(int)

# create a avgerage score feature
df["avg_marks"] =( df[subject_cols].mean(axis=1))*2

# create a attendance score feature
df["attendance_score"] = 1 - (df["absences"] / 15)

# feature selection
features = [
    "avg_marks",
    "previous_marks",
    "study_hours",
    "attendance_score",
    "family_support",
    "school_support"
]

X = df[features]
y = df["pass_fail"]

# Save the cleaned dataset
cleaned_path = "C:\\Users\\kumar\\Academic_Decision_Support_System\\data\\processed\\student_performance_cleaned.csv"


final_df = pd.concat([X, y], axis=1)
final_df.to_csv(cleaned_path, index=False)


print("ML-ready dataset saved to:", cleaned_path)
